currency/updater/requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attr==0.3.2
attrs==23.1.0
Brotli==1.0.9
cattrs==22.2.0
cffi==1.15.1
charset-normalizer==3.1.0
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
motor==3.1.2
motor-stubs==1.7.1
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pycares==4.3.0
pycparser==2.21
pymongo==4.3.3
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
redis==4.5.4
scheduler==0.8.3
typeguard==3.0.2
yarl==1.9.1
